+++ es/cc-context/index.js
@@ -1,4 +1,4 @@
-import _extends from "@babel/runtime/helpers/esm/extends";
+import _extends from "@babel/runtime/helpers/extends";
 
 var _state2, _reducer;
 
@@ -129,8 +129,13 @@ var ccContext = {
     }
   },
   ccKey_ref_: refs,
-  //  key:eventName,  value: Array<{ccKey, identity,  handler}>
+  //  key:eventName,  value: Array<{ccKey, identity,  handlerKey}>
   event_handlers_: {},
+  ccUniqueKey_handlerKeys_: {},
+  // to avoid memory leak, the handlerItem of event_handlers_ just store handlerKey, 
+  // it is a ref that towards ccUniqueKeyEvent_handler_'s key
+  // when component unmounted, it's handler will been removed
+  handlerKey_handler_: {},
   ccKey_option_: {},
   refs: refs,
   info: {