+++ src/core/connect.js
@@ -4,7 +4,7 @@ import { MODULE_DEFAULT } from '../support/constant';
 export default function (ccClassKey, stateToPropMapping, {
   isPropStateModuleMode = false,
   module = MODULE_DEFAULT,
-  extendReactComponent = false,
+  extendInputClass = true,
 } = {}) {
-  return register(ccClassKey, { module, stateToPropMapping, isPropStateModuleMode, extendReactComponent });
+  return register(ccClassKey, { module, stateToPropMapping, isPropStateModuleMode, extendInputClass });
 }